refactor: consolidate duplicate AlgorithmBitmap components
Remove the duplicate AlgorithmBitmap component from features/create/
and update all imports to use the single component in components/.

The components were nearly identical - the only difference was which
service hook they used. The consolidated component uses the SSR-safe
version (useAlgorithmServiceSafe) which works in all contexts.
